Split menaingful locations tests out to separate files
